refactor: move global beta flags to admin form
