+++ src/app/modules/frontend/frontend.service.ts
@@ -1,7 +1,3 @@
-import mongoose from 'mongoose'
-import { okAsync, ResultAsync } from 'neverthrow'
-
-import { SUPPORT_FORM_LINK } from '../../../../shared/constants'
 import { ClientEnvVars } from '../../../../shared/types/core'
 import config from '../../config/config'
 import { captchaConfig } from '../../config/features/captcha.config'
@@ -9,12 +5,6 @@ import { googleAnalyticsConfig } from '../../config/features/google-analytics.co
 import { paymentConfig } from '../../config/features/payment.config'
 import { sentryConfig } from '../../config/features/sentry.config'
 import { spcpMyInfoConfig } from '../../config/features/spcp-myinfo.config'
-import { createLoggerWithLabel } from '../../config/logger'
-import getGlobalBetaModel from '../../models/global_beta.server.model'
-import { DatabaseError } from '../core/core.errors'
-
-const logger = createLoggerWithLabel(module)
-const GlobalBetaModel = getGlobalBetaModel(mongoose)
 
 export const getClientEnvVars = (): ClientEnvVars => {
   return {
@@ -52,25 +42,4 @@ export const getClientEnvVars = (): ClientEnvVars => {
     // TODO (#5826): Toggle to use fetch for submissions instead of axios. Remove once network error is resolved
     useFetchForSubmissions: config.reactMigration.useFetchForSubmissions,
   }
-}
-
-export const getGlobalBetaFlag = (
-  betaFlag: string,
-): ResultAsync<boolean, DatabaseError> => {
-  return ResultAsync.fromPromise(
-    GlobalBetaModel.findFlag(betaFlag),
-    (error) => {
-      logger.error({
-        message: 'Database error when getting global beta flag status',
-        meta: {
-          action: 'findFlag',
-        },
-        error,
-      })
-
-      return new DatabaseError(
-        `Unable to get global beta flag status. If this issue persists, please submit a Support Form at (${SUPPORT_FORM_LINK})`,
-      )
-    },
-  ).andThen((betaFlagDoc) => okAsync(!!betaFlagDoc?.enabled))
 }
